Move route controllers into their own file
